Revert "feat : 상점 판매 기능 구현 / 아이템 리스트 이분화 / UI개선 - 숫자 추가 / 선택문 조건 개선"
This reverts commit b3e42dc74ff30b83cbdf3a56a7b5e417ff628298.
