--- MYproject/1.MainTextAdv.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Drawing;
-using System.Linq;
-using System.Reflection.Emit;
-using System.Text;
-using System.Threading.Tasks;
-
-class TextAdv
-{
-    static void Main(string[] args)
-    {
-        Console.SetWindowSize(105, 43);
-        new Title();
-    }
-}
-
