+++ sparta_dungeon.sln
@@ -3,9 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.8.34322.80
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "sparta_dungeon", "sparta_dungeon\sparta_dungeon.csproj", "{6F6E50CA-BA6E-4320-A7F2-39B453B16547}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MYproject", "MYproject\MYproject.csproj", "{E4CE3BB1-2BAB-43DF-BD34-E42D9DFACDE6}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "sparta_dungeon", "sparta_dungeon\sparta_dungeon.csproj", "{6F6E50CA-BA6E-4320-A7F2-39B453B16547}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -17,10 +15,6 @@ Global
 		{6F6E50CA-BA6E-4320-A7F2-39B453B16547}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6F6E50CA-BA6E-4320-A7F2-39B453B16547}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{6F6E50CA-BA6E-4320-A7F2-39B453B16547}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E4CE3BB1-2BAB-43DF-BD34-E42D9DFACDE6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E4CE3BB1-2BAB-43DF-BD34-E42D9DFACDE6}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E4CE3BB1-2BAB-43DF-BD34-E42D9DFACDE6}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E4CE3BB1-2BAB-43DF-BD34-E42D9DFACDE6}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
